Safari-611.1.10/LayoutTests/js/dom/dom-as-prototype-assignment-exception-expected.txt

Tests to make sure we throw when triggering a custom property with a mismatched this

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testObject.id threw exception TypeError: The Element.id getter can only be used on instances of Element.
PASS testObject.id="foo" threw exception TypeError: The Element.id setter can only be used on instances of Element.
PASS testObject.length threw exception TypeError: The HTMLCollection.length getter can only be used on instances of HTMLCollection.
PASS div.id is 'test'
PASS div.hasOwnProperty('id') is false
PASS div.__proto__.id threw exception TypeError: The Element.id getter can only be used on instances of Element.
PASS successfullyParsed is true

TEST COMPLETE
